webapp/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
qdrant-client==1.6.0
pymongo==4.5.0
neo4j==5.12.0
boto3==1.28.57
requests==2.31.0
psycopg2-binary==2.9.7
temporalio==1.7.0
paho-mqtt==2.1.0